Update TimerMgi.h
changed the setpoint value to an unsigned long from unsigned int

+++ TimerMgi.h
@@ -4,10 +4,10 @@
 class TimerMgi {
     public:
         // Constructor
-        TimerMgi( unsigned int set_point=0, bool elapsed=false );
+        TimerMgi( unsigned long set_point=0, bool elapsed=false );
 
         // Methods
-        void SetSetPoint( unsigned int set_point );
+        void SetSetPoint( unsigned long set_point );
         unsigned int GetSetPoint();
         void TickTimer();
         void ResetTimer();
@@ -15,6 +15,6 @@ class TimerMgi {
 
     private:
         unsigned long curr_time_;
-        unsigned int set_point_;
+        unsigned long set_point_;
         bool elapsed_;
 };
